fix dependencies, update follow refactor
