--- bridger/chains/pangolin/Cargo.toml
@@ -1,43 +1,0 @@
-[package]
-name = "chain-pangolin"
-version = "0.3.3"
-authors = ["Darwinia Network <hello@darwinia.network>"]
-description = "Darwinia bridger"
-repository = "https://github.com/darwinia-network/bridger"
-license = "MIT"
-documentation = "https://docs.rs/bridger"
-homepage = "https://github.com/darwinia-network/bridger"
-include = [
-    "Cargo.toml",
-    "**/*.rs",
-    "README.md",
-    "LICENSE"
-]
-keywords = ["substrate", "darwinia"]
-readme = 'README.md'
-edition = "2018"
-
-[dependencies]
-
-bridge-traits = { path = "../../traits" }
-
-codec                  = { package = "parity-scale-codec", version = "2.0.0" }
-headers-relay          = { git = "https://github.com/darwinia-network/parity-bridges-common.git", branch = "s2s" }
-relay-substrate-client = { git = "https://github.com/darwinia-network/parity-bridges-common.git", branch = "s2s" }
-relay-utils            = { git = "https://github.com/darwinia-network/parity-bridges-common.git", branch = "s2s" }
-
-# Bridge dependencies
-
-drml-primitives         = { git = "https://github.com/darwinia-network/darwinia-common", branch = "s2s" }
-pangolin-runtime        = { git = "https://github.com/darwinia-network/darwinia-common", branch = "s2s" }
-darwinia-ethereum-relay = { git = "https://github.com/darwinia-network/darwinia-common", branch = "s2s" }
-
-
-# Substrate Dependencies
-
-frame-system               = { git = "https://github.com/darwinia-network/substrate.git", branch = "s2s" }
-frame-support              = { git = "https://github.com/darwinia-network/substrate.git", branch = "s2s" }
-pallet-transaction-payment = { git = "https://github.com/darwinia-network/substrate.git", branch = "s2s" }
-sp-core                    = { git = "https://github.com/darwinia-network/substrate.git", branch = "s2s" }
-sp-keyring                 = { git = "https://github.com/darwinia-network/substrate.git", branch = "s2s" }
-sp-runtime                 = { git = "https://github.com/darwinia-network/substrate.git", branch = "s2s" }
