+++ bridger/components/client-millau/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "component-pangolin"
+name = "component-millau"
 version = "0.3.3"
 authors = ["Darwinia Network <hello@darwinia.network>"]
 description = "Darwinia bridger"
@@ -26,19 +26,16 @@ headers-relay          = { git = "https://github.com/darwinia-network/parity-bri
 relay-substrate-client = { git = "https://github.com/darwinia-network/parity-bridges-common.git", branch = "main" }
 relay-utils            = { git = "https://github.com/darwinia-network/parity-bridges-common.git", branch = "main" }
 
-# Bridge dependencies
+# Supported Chains
 
-drml-primitives         = { git = "https://github.com/darwinia-network/darwinia-common", branch = "s2s" }
-pangolin-runtime        = { git = "https://github.com/darwinia-network/darwinia-common", branch = "s2s" }
-darwinia-ethereum-relay = { git = "https://github.com/darwinia-network/darwinia-common", branch = "s2s" }
-
+millau-runtime      = { git = "https://github.com/darwinia-network/darwinia-launch.git", branch = "s2s" }
+millau-primitives   = { git = "https://github.com/darwinia-network/darwinia-launch.git", branch = "s2s" }
 
 # Substrate Dependencies
 
-frame-system               = { git = "https://github.com/darwinia-network/substrate.git", branch = "main" }
 frame-support              = { git = "https://github.com/darwinia-network/substrate.git", branch = "main" }
+frame-system               = { git = "https://github.com/darwinia-network/substrate.git", branch = "main" }
 pallet-transaction-payment = { git = "https://github.com/darwinia-network/substrate.git", branch = "main" }
 sp-core                    = { git = "https://github.com/darwinia-network/substrate.git", branch = "main" }
 sp-keyring                 = { git = "https://github.com/darwinia-network/substrate.git", branch = "main" }
 sp-runtime                 = { git = "https://github.com/darwinia-network/substrate.git", branch = "main" }
-
